+++ site/routes_examples.go
@@ -106,6 +106,9 @@ func setupExamples(ctx context.Context, router chi.Router, signals sessions.Stor
 				{ID: "click_outside"},
 				{ID: "timing"},
 				{ID: "aliased_data_attributes"},
+				{ID: "merge_fragment"},
+				{ID: "key_casing"},
+				{ID: "signals_ifmissing_onload"},
 			},
 		},
 	}
@@ -201,6 +204,7 @@ func setupExamples(ctx context.Context, router chi.Router, signals sessions.Stor
 			setupExamplesToggleVisibility(examplesRouter),
 			setupExamplesFormData(examplesRouter),
 			setupExamplesCustomValidity(examplesRouter),
+			setupExamplesMergeFragment(examplesRouter),
 		); err != nil {
 			panic(fmt.Sprintf("error setting up examples routes: %s", err))
 		}